Remove dependencies on #playerContainer id
paella.PlayerBase.initialize takes a playerId as the first argument.
This allows changing the id and in theory would allow multiple players
in one document. This id is used in a lot of places to construct unique
ids, but there are some places that expect a container with the id
`playerContainer`.

This tries to fix this. setupTimeOnly and setupTimeImageOverlay may
need further attention as they completely ignore the existing
paella.DomNode facilities.