Fixes Matomo Media Analytics Tool title
The media title was only sending to MMAT when was integred in Opencast,
but no with Paella player as standalone. This fix enables the title
to be send to Matomo MAT and ads a fallback option "html_title" when is
requiered to get the html title instead of the name of the video.